[优化]：整理高亮规则文件
