+++ app/src/main/java/io/legado/app/ui/book/read/config/MoreConfigDialog.kt
@@ -25,6 +25,7 @@ import io.legado.app.lib.theme.primaryColor
 import io.legado.app.model.ReadBook
 import io.legado.app.ui.book.read.ReadBookActivity
 import io.legado.app.ui.book.read.TextMenuConfigDialog
+import io.legado.app.ui.book.read.config.highlight.HighlightRuleConfigDialog
 import io.legado.app.ui.book.read.page.provider.ChapterProvider
 import io.legado.app.ui.widget.number.NumberPickerDialog
 import io.legado.app.utils.canvasrecorder.CanvasRecorderFactory
