+++ app/src/main/java/io/legado/app/ui/book/read/config/highlight/HighlightRuleGroupManageDialog.kt
@@ -1,4 +1,4 @@
-package io.legado.app.ui.book.read.config
+package io.legado.app.ui.book.read.config.highlight
 
 import android.graphics.PorterDuff
 import android.os.Bundle
@@ -18,8 +18,6 @@ import io.legado.app.base.BaseDialogFragment
 import io.legado.app.base.adapter.ItemViewHolder
 import io.legado.app.base.adapter.RecyclerAdapter
 import io.legado.app.constant.EventBus
-import io.legado.app.ui.book.read.config.highlight.HighlightRuleGroupStore
-import io.legado.app.ui.book.read.config.highlight.HighlightRuleRepository
 import io.legado.app.databinding.DialogHighlightRuleGroupManageBinding
 import io.legado.app.lib.dialogs.alert
 import io.legado.app.lib.theme.accentColor
