+++ app/src/main/java/io/legado/app/ui/book/read/config/highlight/HighlightRulePreview.kt
@@ -1,20 +1,8 @@
-package io.legado.app.ui.book.read.config
+package io.legado.app.ui.book.read.config.highlight
 
 import android.text.SpannableStringBuilder
 import android.text.Spanned
 import android.text.style.ForegroundColorSpan
-import io.legado.app.ui.book.read.config.BgColorSpan
-import io.legado.app.ui.book.read.config.BgImageSpan
-import io.legado.app.ui.book.read.config.DoubleUnderlineSpan
-import io.legado.app.ui.book.read.config.SvgUnderlineSpan
-import io.legado.app.ui.book.read.config.SolidUnderlineSpan
-import io.legado.app.ui.book.read.config.DashUnderlineSpan
-import io.legado.app.ui.book.read.config.WaveUnderlineSpan
-import io.legado.app.ui.book.read.config.StrikeThroughSpan
-import io.legado.app.ui.book.read.config.ItalicTextSpan
-import io.legado.app.ui.book.read.config.BoxTextSpan
-import io.legado.app.ui.book.read.config.highlight.HighlightRule
-import io.legado.app.ui.book.read.config.highlight.HighlightRuleStyle
 
 /**
  * 高亮规则配置页的预览文本构建器。
